// src/app/models/email/email-message.model.ts
export interface EmailMessageModel {

    rowCount: number;

    mailFrom: string;

    toRecipients: string;

    subject: string;

    body: string;

    attachmentPath: string;

    ccRecipients: string;

    status: string;

    header: string;

    footer: string;

    bodyAsHtml: string;

    error: string

    emailSetting: EmailSettingModel;
}


export interface EmailSettingModel {
    smtpServer: string;
    port: number;
    userName: string;
    password: string;
}